Add files via upload
This zip file contains all the assets used in the game, after extracting it the game should work flawlessly by uploading running the objserver file, then the odbjclient file. If you want to play single player you have to download the single player file seperate from this one and still download this zip file to extract, so you have all the assets and the lables file. Furthermore if the directory for the game assets isn't working they are all set in the labels file so that is where to look.